Move queue.ts under src/main
