+++ src/server.ts
@@ -17,12 +17,12 @@ import { ConnectionOptions } from 'tls'
 import { defaultConfig } from './config'
 import { DB } from './db'
 import { EventsProcessor } from './ingestion/process-event'
+import { startQueue } from './main/queue'
 import { startSchedule } from './main/services/schedule'
 import { status } from './status'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from './types'
 import { createPostgresPool, createRedis, delay, UUIDT } from './utils'
 import { startFastifyInstance, stopFastifyInstance } from './web/server'
-import { startQueue } from './worker/queue'
 
 const { version } = require('../package.json')
